refactor WebsocketServer
